showing data in table
