+++ utility/index.ts
@@ -1,0 +1,4 @@
+export const isObject = (obj: any) =>
+  typeof obj === "object" && !Array.isArray(obj) && obj !== null;
+
+export const isBigNumber = (obj: any) => obj?.['type'] === 'BigNumber'
